requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy
scikit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
openai==1.3.7
deep-translator==1.11.4
google-generativeai==0.3.2
geopy==2.4.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
